requirements_full.txt
-i https://pypi.org/simple
appdirs==1.4.4
black==21.4b2
click==8.1.7 ; python_version >= '3.7'
exceptiongroup==1.2.1 ; python_version < '3.11'
iniconfig==2.0.0 ; python_version >= '3.7'
markdown-it-py==3.0.0 ; python_version >= '3.8'
mdurl==0.1.2 ; python_version >= '3.7'
mypy[cli]==1.10.1
mypy-extensions==1.0.0 ; python_version >= '3.5'
packaging==24.1 ; python_version >= '3.8'
pathspec==0.12.1 ; python_version >= '3.8'
pluggy==1.5.0 ; python_version >= '3.8'
pygments==2.18.0 ; python_version >= '3.8'
pytest==8.2.2
python-dotenv==1.0.1 ; python_version >= '3.8'
regex==2024.5.15 ; python_version >= '3.8'
rich==13.7.1 ; python_full_version >= '3.7.0'
shellingham==1.5.4 ; python_version >= '3.7'
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1 ; python_version < '3.11'
typer==0.12.3
typing-extensions==4.12.2 ; python_version >= '3.8'
annotated-types==0.7.0 ; python_version >= '3.8'
attrs==23.2.0 ; python_version >= '3.7'
beautifulsoup4==4.13.0b2
brotli==1.1.0
cachetools==5.3.3 ; python_version >= '3.7'
certifi==2024.6.2 ; python_version >= '3.6'
chardet==3.0.4
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
google-api-core[grpc]==2.19.1 ; python_version >= '3.7'
google-auth==2.30.0 ; python_version >= '3.7'
google-cloud-core==2.4.1 ; python_version >= '3.7'
google-cloud-pubsub==2.21.5
google-cloud-storage==2.17.0
google-crc32c==1.5.0 ; python_version >= '3.7'
google-resumable-media==2.7.1 ; python_version >= '3.7'
googleapis-common-protos==1.63.2 ; python_version >= '3.7'
grpc-google-iam-v1==0.13.1 ; python_version >= '3.7'
grpcio==1.65.0rc1 ; python_version >= '3.8'
grpcio-status==1.65.0rc1 ; python_version >= '3.8'
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.10 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
price-parser==0.3.4
proto-plus==1.24.0 ; python_version >= '3.7'
protobuf==5.27.1 ; python_version >= '3.8'
psycopg2-binary==2.9.9
pyasn1==0.6.0 ; python_version >= '3.8'
pyasn1-modules==0.4.0 ; python_version >= '3.8'
pydantic==2.7.4 ; python_version >= '3.8'
pydantic-core==2.18.4 ; python_version >= '3.8'
pydantic-settings==2.3.4
pydash==8.0.1
requests==2.32.3
rfc3986==1.5.0
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
soupsieve==2.5 ; python_version >= '3.8'
structlog==21.5.0
tqdm==4.66.4
types-requests==2.32.0.20240622
unidecode==1.3.8
urllib3==2.2.2 ; python_version >= '3.8'
